ipc/glue/test/browser/test_android_audio_on.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Audio decoder in Utility process (AAC)</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script type="text/javascript">
SimpleTest.waitForExplicitFinish();

(async function() {
  // FIXME: Remove all after Bug 1934009
  for (let src of [
    "small-shot.m4a",
  ]) {
    try {
      await runMochitestUtilityAudio(src, { expectUtility: "", expectDecoder: "android decoder", expectContent: false, expectJava: true });
    } catch (ex) {
      ok(false, `Failure for ${src} ${ex}`);
    }
  }

  SimpleTest.finish();
})();
</script>
</pre>
</body>
</html>
